+++ Code/Core/FileIO/MemoryStream.h
@@ -18,7 +18,7 @@ public:
     // memory stream specific functions
     inline const void * GetData() const { return (void *)m_Begin; }
     inline void *       GetDataMutable() { return (void *)m_Begin; }
-    inline size_t       GetSize() const { return ( m_End - m_Begin ); }
+    inline size_t       GetSize() const { return (size_t)( m_End - m_Begin ); }
     void                Reset();
 
     uint64_t            WriteBuffer( IOStream & stream, uint64_t bytesToWrite );
